requirements.txt
streamlit>=1.34
pymupdf>=1.24
pillow>=10.0
requests>=2.31
